[add] Upd001_トークテーマ投稿の実装
・単項目バリデーションの実装
・TODO：テストクラスの実装

+++ eternal-app/src/main/java/com/eternal/web/repository/CategoryRepository.java
@@ -3,7 +3,6 @@
  */
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.CategoryEntity;
@@ -14,6 +13,6 @@ import com.eternal.web.entity.CategoryEntity;
  * @author taiki0304
  */
 @Repository
-public interface CategoryRepository extends JpaRepository<CategoryEntity, BigDecimal> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
 
 }
