+++ eternal-app/src/main/java/com/eternal/web/repository/TalkHistoryRepository.java
@@ -3,7 +3,6 @@
  */
 package com.eternal.web.repository;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.eternal.web.entity.TalkHistoryEntity;
@@ -14,6 +13,6 @@ import com.eternal.web.entity.TalkHistoryEntity;
  * @author taiki0304
  */
 @Repository
-public interface TalkHistoryRepository extends JpaRepository<TalkHistoryEntity, BigDecimal> {
+public interface TalkHistoryRepository extends JpaRepository<TalkHistoryEntity, Long> {
 
 }
